# backend-express/swagger.yaml
openapi: 3.0.0
info:
  title: Kopi Haus Dashboard API
  version: 1.0.0
  description: API documentation for the Kopi Haus Dashboard backend application.
servers:
  - url: http://localhost:5001/api
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: User registered successfully
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Server error
...
    Shift:
      type: object
      required:
        - register_id
        - opened_by
        - opened_at
      properties:
        _id:
          type: string
          readOnly: true
          example: 60d5ec49f8c7a40015a7b8c12
        register_id:
          type: string
          example: 60d5ec49f8c7a40015a7b8c4
        opened_by:
          type: string
          example: 60d5ec49f8c7a40015a7b8c13
        closed_by:
          type: string
          nullable: true
          example: 60d5ec49f8c7a40015a7b8c14
        opened_at:
          type: string
          format: date-time
          example: '2023-10-26T10:00:00.000Z'
        closed_at:
          type: string
          format: date-time
          nullable: true
          example: '2023-10-26T18:00:00.000Z'
        opening_cash:
          type: number
          format: float
          example: 1000000
        closing_cash:
          type: number
          format: float
          nullable: true
          example: 2500000
        created_at:
          type: string
          format: date-time
          readOnly: true
          example: '2023-10-26T10:00:00.000Z'
...
    StockAdjustment:
      type: object
      required:
        - outlet_id
        - adjusted_by
        - adjusted_at
      properties:
        _id:
          type: string
          readOnly: true
          example: 60d5ec49f8c7a40015a7b8c22
        outlet_id:
          type: string
          description: ID of the outlet where the stock adjustment was made
          example: 60d5ec49f8c7a40015a7b8c1
        adjusted_by:
          type: string
          description: ID of the user who made the stock adjustment
          example: 60d5ec49f8c7a40015a7b8c13
        adjusted_at:
          type: string
          format: date-time
          description: Timestamp when the stock adjustment was made
          example: '2023-10-26T12:00:00.000Z'
        reason:
          type: string
          nullable: true
          description: Reason for the stock adjustment
          example: Inventory count discrepancy
        created_at:
          type: string
          format: date-time
          readOnly: true
          example: '2023-10-26T12:00:00.000Z'
    StockAdjustmentItem:
      type: object
      required:
        - stock_adjustment_id
        - product_id
        - delta_qty
      properties:
        _id:
          type: string
          readOnly: true
          example: 60d5ec49f8c7a40015a7b8c23
        stock_adjustment_id:
          type: string
          description: ID of the associated stock adjustment
          example: 60d5ec49f8c7a40015a7b8c22
        product_id:
          type: string
          description: ID of the product being adjusted
          example: 60d5ec49f8c7a40015a7b8c6
        delta_qty:
          type: number
          format: integer
          description: Change in quantity (positive for increase, negative for decrease)
          example: 5
        created_at:
          type: string
          format: date-time
          readOnly: true
          example: '2023-10-26T12:00:00.000Z'
    StockMovement:
      type: object
      required:
        - product_id
        - outlet_id
        - source_type
        - source_id
        - balance_after
        - moved_at
      properties:
        _id:
          type: string
          readOnly: true
          example: 60d5ec49f8c7a40015a7b8c24
        product_id:
          type: string
          description: ID of the product involved in the movement
          example: 60d5ec49f8c7a40015a7b8c6
        outlet_id:
          type: string
          description: ID of the outlet where the movement occurred
          example: 60d5ec49f8c7a40015a7b8c1
        source_type:
          type: string
          enum: [purchase, sale, return_sales, return_purchase, stock_adjustment, stock_adjustment_revert, return_sales_delete_revert, return_purchase_delete_revert]
          description: Type of event that caused the stock movement
          example: purchase
        source_id:
          type: string
          description: ID of the related document (e.g., Purchase, Sale, Return, StockAdjustment)
          example: 60d5ec49f8c7a40015a7b8c15
        qty_in:
          type: number
          format: integer
          description: Quantity moved into stock
          example: 5
        qty_out:
          type: number
          format: integer
          description: Quantity moved out of stock
          example: 0
        balance_after:
          type: number
          format: integer
          description: Stock quantity after the movement
          example: 105
        moved_at:
          type: string
          format: date-time
          description: Timestamp of the stock movement
          example: '2023-10-26T13:00:00.000Z'
...
paths:
...
  /shifts:
    get:
      summary: Get all shifts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shift'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /shifts/open:
    post:
      summary: Open a new shift
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - register_id
                - opened_by
                - opening_cash
              properties:
                register_id:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c4
                opened_by:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c13
                opening_cash:
                  type: number
                  format: float
                  example: 1000000
      responses:
        '201':
          description: Shift opened successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '400':
          description: Invalid input or an open shift already exists for this register
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /shifts/close/{id}:
    patch:
      summary: Close an existing shift
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the shift to close
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - closed_by
                - closing_cash
              properties:
                closed_by:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c14
                closing_cash:
                  type: number
                  format: float
                  example: 2500000
      responses:
        '200':
          description: Shift closed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '400':
          description: Invalid input or shift already closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Shift not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /shifts/{id}:
    delete:
      summary: Delete a shift by ID (only if not closed)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the shift to delete
      responses:
        '200':
          description: Shift deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Cannot delete an open shift
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Shift not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /stock-adjustments:
    get:
      summary: Get all stock adjustments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of stock adjustments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockAdjustment'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new stock adjustment with items
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - outlet_id
                - adjusted_by
                - items
              properties:
                outlet_id:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c1
                adjusted_by:
                  type: string
                  example: 60d5ec49f8c7a40015a7b8c13
                reason:
                  type: string
                  nullable: true
                  example: Damaged stock
                items:
                  type: array
                  items:
                    type: object
                    required:
                      - product_id
                      - delta_qty
                    properties:
                      product_id:
                        type: string
                        example: 60d5ec49f8c7a40015a7b8c6
                      delta_qty:
                        type: number
                        format: integer
                        example: -2
      responses:
        '201':
          description: Stock adjustment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  stockAdjustment:
                    $ref: '#/components/schemas/StockAdjustment'
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/StockAdjustmentItem'
        '400':
          description: Invalid input, outlet ID, user ID, product ID, or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stock-adjustments/{id}:
    get:
      summary: Get a stock adjustment by ID with its items
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the stock adjustment to retrieve
      responses:
        '200':
          description: Stock adjustment data with items
          content:
            application/json:
              schema:
                type: object
                properties:
                  stockAdjustment:
                    $ref: '#/components/schemas/StockAdjustment'
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/StockAdjustmentItem'
        '404':
          description: Stock adjustment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a stock adjustment by ID (and its items)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the stock adjustment to delete
      responses:
        '200':
          description: Stock adjustment and its items deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Stock adjustment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stock-movements:
    get:
      summary: Get all stock movements
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of stock movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockMovement'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stock-movements/{id}:
    get:
      summary: Get a stock movement by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the stock movement to retrieve
      responses:
        '200':
          description: Stock movement data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockMovement'
        '404':
          description: Stock movement not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...